+++ frontend/lost/src/components/SIA/lost-sia/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Confirm label in LabelInput by click on the respective label
 - Configure name of the default Label by a prop 
 - Provide method to reset canvas zoom
+- It is now possible to define custom label colors via canvas *possibleLabels* props 
 
 ### Fixed
 - Fixed jumping camera when zooming into the image